pyproject.toml
[project]
name = "uat-automation-framework"
version = "0.1.0"
description = "UAT Automation Framework with Python, Behave, and Selenium"
authors = [
    {name = "Bruno Almeida"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "selenium (>=4.39.0,<5.0.0)",
    "behave (>=1.3.3,<2.0.0)",
    "allure-behave (>=2.15.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
